Separate CLI and library functions of eeprom utils
