+++ dockers/docker-platform-monitor/Dockerfile.j2
@@ -62,7 +62,7 @@ RUN pip3 install thrift==0.13.0
 RUN pip3 install requests
 
 # Nexthop platform vendors' sonic_platform packages import these Python libraries
-RUN pip3 install python-dateutil==2.9.0.post0 click
+RUN pip3 install python-dateutil==2.9.0.post0
 
 # We install the libpci module in order to be able to do PCI transactions
 RUN pip3 install libpci
