+++ platform/broadcom/sonic-platform-modules-nexthop/common/utils/idprom
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from nexthop_utils.eeprom_utils import cli
+from nexthop_utils.eeprom_utils_cli import cli
 
 if __name__ == "__main__":
     cli()
